order/api/views.py
from order.models import Order, Status
from rest_framework import viewsets
from order.api.serializers import StatusSerializer, OrderSerializer


class StatusViewSet(viewsets.ModelViewSet):
    queryset = Status.objects.all()
    serializer_class = StatusSerializer


class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all()
    serializer_class = OrderSerializer
